refactor(채팅): 불필요한 주석 제거
