+++ backend/src/main/java/skkuchin/service/config/chat/KafkaConfig.java
@@ -15,6 +15,7 @@ import java.util.Map;
 
 @Configuration
 @EnableKafka
+//bootstrapServers, groupId 수정해야함
 public class KafkaConfig {
